- updated to SharpDX 2.2.0 - removed old SharpDX-Win8 assemblies

+++ RenderSystems/ANX.Framework.Windows.DX10/GraphicsDeviceWindowsDX10.cs
@@ -633,7 +633,7 @@ namespace ANX.RenderSystem.Windows.DX10
 
                 //TODO: handle format
 
-                swapChain.ResizeBuffers(swapChain.Description.BufferCount, presentationParameters.BackBufferWidth, presentationParameters.BackBufferHeight, Format.R8G8B8A8_UNorm, (int)swapChain.Description.Flags);
+                swapChain.ResizeBuffers(swapChain.Description.BufferCount, presentationParameters.BackBufferWidth, presentationParameters.BackBufferHeight, Format.R8G8B8A8_UNorm, swapChain.Description.Flags);
 
                 backBuffer = SharpDX.Direct3D10.Texture2D.FromSwapChain<SharpDX.Direct3D10.Texture2D>(swapChain, 0);
                 renderView = new RenderTargetView(device, backBuffer);
